Add Japanese R groups
Annual events are held by "Japan.R" and monthly events are held in various cities. Some groups may no longer be active. "Tokyo.R" is the largest and most active group. Many groups are active on Twitter and migrating from ATND to the Connpass system to organise events. Events are organised in Japanese but many organisers can be contacted in English.